fix(ci): move bindings so they are easier to pack
