+++ .github/workflows/ci.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - main
-      - 'releases/*'
+      - 'release/*'
 jobs:
   lint:
     runs-on: ubuntu-latest
