+++ binding.gyp
@@ -2,7 +2,7 @@
   "targets": [
     {
       "target_name": "cpu_profiler",
-      "sources": [ "src/bindings/cpu_profiler.cc" ],
+      "sources": [ "bindings/cpu_profiler.cc" ],
       "defines": ["PROFILER_FORMAT=FORMAT_SAMPLED"],
       'include_dirs': [
         '<!(node -e "require(\'nan\')")'
@@ -10,7 +10,7 @@
     },
     # {
     #   "target_name": "cpu_profiler_format_benchmark",
-    #   "sources": [ "src/bindings/cpu_profiler.cc" ],
+    #   "sources": [ "bindings/cpu_profiler.cc" ],
     #   "defines": ["FORMAT_BENCHMARK=1"],
     #   'include_dirs': [
     #     '<!(node -e "require(\'nan\')")'
